Woodstocks/requirements.txt
tweepy>=4.0
yfinance>=0.2
textblob>=0.17
scikit-learn>=1.0
xgboost>=1.5
pandas>=1.3
matplotlib>=3.5
seaborn>=0.11
plotly>=5.0
python-dotenv>=0.19
joblib>=1.0
nltk>=3.7